refactor: Elective Grid Item
